refactor(zoho): structure oauth wizard module
